Add files via upload
In this Python code used edge tts (text-to-speech) for responce using american jessaneural as a default voice which gives calm responses adjust the voice through ssml to make it some human like and used stt for text to speech , Here whever we speak it will get converted into text and the responce will get converted into speech and get saved in .wav file. It uses NLP based phi model-1.5 whihc is good for CPU and gives a 100 percent correct and actual responce.